third_party/WebKit/LayoutTests/http/tests/preload/reuse_sri.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
success = function() {}
var t = async_test('Ensure that elements with SRI reuse their preload');
window.addEventListener('load', t.step_func(function() {
  var seen = {};
  performance.getEntriesByType("resource").forEach(function(element) {
    var name = element.name;
    assert_false(!!seen[name], name + " was loaded more than once");
    seen[name] = true;
  });
  t.done();
}));
</script>
<link rel="stylesheet" href="../security/subresourceIntegrity/style-2-of-3.css" integrity="sha384-j1bGpDxu0bZAb4jBJzciDWuVUuOTXlLvB9WNKDh7Gik9ygF6CctjgYISUwShHmFM"></link>
<script src="../security/subresourceIntegrity/call-success.js" integrity="sha256-B0/62fJSJFrdjEFR9ba04m/D+LHQ+zG6PGcaR0Trpxg="></script>
